http://ftp.ksu.edu.tw/FTP/FreeBSD/distfiles/xscreensaver-4.20.tar.gz -rw-r--r-- 1 zblaxell zblaxell 4437919 Mar 14  2005 xscreensaver-4.20.tar.gz 304da3cfa2c02aa4cedca85b48005b4163d94452  xscreensaver-4.20.tar.gz

+++ hacks/intermomentary.c
@@ -282,7 +282,7 @@ inline void moverender_rider(Display *dpy, Window window, GC fgc, struct field *
     double cs, cv;
 
     /* add velocity to theta */
-    rid->t = fmodf((rid->t + rid->vt + M_PI), (2 * M_PI)) - M_PI;
+    rid->t = fmod((rid->t + rid->vt + M_PI), (2 * M_PI)) - M_PI;
     
     rid->vt += frand(0.002) - 0.001;
